Refactoring and modernizing the nodeJS SDK
Broke up several functions to their own files.
Created a semantic file structure for utils of the same family (encode, parse)
Updated old variable declarations from 'var' to 'let'
