+++ lib/auth_token.js
@@ -4,7 +4,7 @@
  */
 
 const crypto = require('crypto');
-const { smart_escape } = require('./utils/encoding/smart_escape');
+const smart_escape = require('./utils/encoding/smart_escape');
 
 const unsafe = /([ "#%&'/:;<=>?@[\]^`{|}~]+)/g;
