+++ lib/preloaded_file.js
@@ -1,4 +1,4 @@
-var PRELOADED_CLOUDINARY_PATH, config, utils;
+let PRELOADED_CLOUDINARY_PATH, config, utils;
 
 utils = require("./utils");
 
@@ -8,7 +8,7 @@ PRELOADED_CLOUDINARY_PATH = /^([^\/]+)\/([^\/]+)\/v(\d+)\/([^#]+)#([^\/]+)$/;
 
 class PreloadedFile {
   constructor(file_info) {
-    var matches, public_id_and_format;
+    let matches, public_id_and_format;
     matches = file_info.match(PRELOADED_CLOUDINARY_PATH);
     if (!matches) {
       throw "Invalid preloaded file info";
@@ -24,7 +24,7 @@ class PreloadedFile {
   }
 
   is_valid() {
-    var expected_signature;
+    let expected_signature;
     expected_signature = utils.api_sign_request({
       public_id: this.public_id,
       version: this.version,
@@ -33,7 +33,7 @@ class PreloadedFile {
   }
 
   static split_format(identifier) {
-    var format, last_dot, public_id;
+    let format, last_dot, public_id;
     last_dot = identifier.lastIndexOf(".");
     if (last_dot === -1) {
       return [identifier, null];
@@ -52,7 +52,7 @@ class PreloadedFile {
   }
 
   toJSON() {
-    var result = {};
+    let result = {};
     Object.getOwnPropertyNames(this).forEach((key) => {
       let val = this[key];
       if (typeof val !== 'function') {
