+++ lib/uploader.js
@@ -366,7 +366,7 @@ function call_api(action, callback, options, get_params) {
   let boundary = utils.random_public_id();
   let errorRaised = false;
   let handle_response = function (res) {
-    // var buffer;
+    // let buffer;
     if (errorRaised) {
 
       // Already reported
@@ -424,7 +424,7 @@ function call_api(action, callback, options, get_params) {
 }
 
 function post(url, post_data, boundary, file, callback, options) {
-  var file_header;
+  let file_header;
   let finish_buffer = Buffer.from("--" + boundary + "--", 'ascii');
   if ((file != null) || options.stream) {
     let filename = options.stream ? "file" : basename(file);
