+++ lib/utils/defaults.js
@@ -61,6 +61,75 @@ const LAYER_KEYWORD_PARAMS = {
 };
 
 
+const TRANSFORMATION_PARAMS = [
+  'angle',
+  'aspect_ratio',
+  'audio_codec',
+  'audio_frequency',
+  'background',
+  'bit_rate',
+  'border',
+  'color',
+  'color_space',
+  'crop',
+  'default_image',
+  'delay',
+  'density',
+  'dpr',
+  'duration',
+  'effect',
+  'end_offset',
+  'fetch_format',
+  'flags',
+  'fps',
+  'gravity',
+  'height',
+  'if',
+  'keyframe_interval',
+  'offset',
+  'opacity',
+  'overlay',
+  'page',
+  'prefix',
+  'quality',
+  'radius',
+  'raw_transformation',
+  'responsive_width',
+  'size',
+  'start_offset',
+  'streaming_profile',
+  'transformation',
+  'underlay',
+  'variables',
+  'video_codec',
+  'video_sampling',
+  'width',
+  'x',
+  'y',
+  'zoom', // + any key that starts with '$'
+];
+
+let SIMPLE_PARAMS = [
+  ["audio_codec", "ac"],
+  ["audio_frequency", "af"],
+  ["bit_rate", 'br'],
+  ["color_space", "cs"],
+  ["default_image", "d"],
+  ["delay", "dl"],
+  ["density", "dn"],
+  ["duration", "du"],
+  ["end_offset", "eo"],
+  ["fetch_format", "f"],
+  ["gravity", "g"],
+  ["page", "pg"],
+  ["prefix", "p"],
+  ["start_offset", "so"],
+  ["streaming_profile", "sp"],
+  ["video_codec", "vc"],
+  ["video_sampling", "vs"],
+];
+
+
 module.exports = {
   DEFAULT_RESPONSIVE_WIDTH_TRANSFORMATION,
   DEFAULT_POSTER_OPTIONS,
@@ -68,4 +137,6 @@ module.exports = {
   CONDITIONAL_OPERATORS,
   PREDEFINED_VARS,
   LAYER_KEYWORD_PARAMS,
+  TRANSFORMATION_PARAMS,
+  SIMPLE_PARAMS,
 };
