+++ lib/utils/encoding/smart_escape.js
@@ -9,4 +9,4 @@ function smart_escape(string, unsafe = /([^a-zA-Z0-9_.\-\/:]+)/g) {
   });
 }
 
-module.exports.smart_escape = smart_escape;
+module.exports = smart_escape;
