+++ lib/utils/entries.js
@@ -1,5 +1,5 @@
 module.exports = Object.entries ? Object.entries : function (obj) {
-  var ownProps = Object.keys(obj),
+  let ownProps = Object.keys(obj),
     i = ownProps.length,
     resArray = new Array(i); // preallocate the Array
   while (i--) {
