+++ lib/utils/index.js
@@ -30,9 +30,11 @@ const isObject = require("lodash/isObject");
 const isString = require("lodash/isString");
 const isUndefined = require("lodash/isUndefined");
 
-const { smart_escape } = require("./encoding/smart_escape");
-const { consumeOption } = require('./parsing/consumeOption');
-const { toArray } = require('./parsing/toArray');
+const smart_escape = require("./encoding/smart_escape");
+const consumeOption = require('./parsing/consumeOption');
+const toArray = require('./parsing/toArray');
+let { base64EncodeURL } = require('./encoding/base64EncodeURL');
+const encodeDoubleArray = require('./encoding/encodeDoubleArray');
 
 const config = require("../config");
 const generate_token = require("../auth_token");
@@ -42,23 +44,6 @@ const ensurePresenceOf = require('./ensurePresenceOf');
 const ensureOption = require('./ensureOption').defaults(config());
 const entries = require('./entries');
 const isRemoteUrl = require('./isRemoteUrl');
-
-module.exports = {
-  at,
-  clone,
-  extend,
-  filter,
-  includes,
-  isArray,
-  isEmpty,
-  isNumber,
-  isObject,
-  isRemoteUrl,
-  isString,
-  isUndefined,
-  keys: source => Object.keys(source),
-  ensurePresenceOf,
-};
 
 exports = module.exports;
 const utils = module.exports;
@@ -96,6 +81,8 @@ const {
   CONDITIONAL_OPERATORS,
   PREDEFINED_VARS,
   LAYER_KEYWORD_PARAMS,
+  TRANSFORMATION_PARAMS,
+  SIMPLE_PARAMS,
 } = require('./defaults');
 
 function textStyle(layer) {
@@ -193,7 +180,7 @@ function process_if(ifValue) {
  * @return {string} layer transformation string
  */
 function process_layer(layer) {
-  var result = '';
+  let result = '';
   if (isPlainObject(layer)) {
     if (layer.resource_type === "fetch" || (layer.url != null)) {
       result = `fetch:${base64EncodeURL(layer.url)}`;
@@ -282,23 +269,6 @@ function process_radius(radius) {
   return radius.map(normalize_expression).join(':');
 }
 
-function base64EncodeURL(sourceUrl) {
-  try {
-    sourceUrl = decodeURI(sourceUrl);
-  } catch (error) {
-    // ignore errors
-  }
-  sourceUrl = encodeURI(sourceUrl);
-  return base64Encode(sourceUrl);
-}
-
-function base64Encode(input) {
-  if (!(input instanceof Buffer)) {
-    input = Buffer.from(String(input), 'binary');
-  }
-  return input.toString('base64');
-}
-
 function build_upload_params(options) {
   let params = {
     access_mode: options.access_mode,
@@ -336,20 +306,6 @@ function build_upload_params(options) {
   return utils.updateable_resource_params(options, params);
 }
 
-/**
- * Serialize an array of arrays into a string
- * @param {[]|[[]]} array - An array of arrays.
- *                          If the first element is not an array the argument is wrapped in an array.
- * @returns {string} A string representation of the arrays.
- */
-
-function encode_double_array(array) {
-  array = toArray(array);
-  if (!isArray(array[0])) {
-    array = [array];
-  }
-  return array.map(e => toArray(e).join(",")).join("|");
-}
 function encode_key_value(arg) {
   if (!isObject(arg)) { return arg; }
   return entries(arg).map(([k, v]) => `${k}=${v}`).join('|');
@@ -387,53 +343,6 @@ function build_custom_headers(headers) {
       return headers;
   }
 }
-const TRANSFORMATION_PARAMS = [
-  'angle',
-  'aspect_ratio',
-  'audio_codec',
-  'audio_frequency',
-  'background',
-  'bit_rate',
-  'border',
-  'color',
-  'color_space',
-  'crop',
-  'default_image',
-  'delay',
-  'density',
-  'dpr',
-  'duration',
-  'effect',
-  'end_offset',
-  'fetch_format',
-  'flags',
-  'fps',
-  'gravity',
-  'height',
-  'if',
-  'keyframe_interval',
-  'offset',
-  'opacity',
-  'overlay',
-  'page',
-  'prefix',
-  'quality',
-  'radius',
-  'raw_transformation',
-  'responsive_width',
-  'size',
-  'start_offset',
-  'streaming_profile',
-  'transformation',
-  'underlay',
-  'variables',
-  'video_codec',
-  'video_sampling',
-  'width',
-  'x',
-  'y',
-  'zoom', // + any key that starts with '$'
-];
 
 function generate_transformation_string(options) {
   if (utils.isString(options)) {
@@ -530,27 +439,8 @@ function generate_transformation_string(options) {
     y: normalize_expression(consumeOption(options, "y")),
     z: normalize_expression(consumeOption(options, "zoom")),
   };
-  let simple_params = [
-    ["audio_codec", "ac"],
-    ["audio_frequency", "af"],
-    ["bit_rate", 'br'],
-    ["color_space", "cs"],
-    ["default_image", "d"],
-    ["delay", "dl"],
-    ["density", "dn"],
-    ["duration", "du"],
-    ["end_offset", "eo"],
-    ["fetch_format", "f"],
-    ["gravity", "g"],
-    ["page", "pg"],
-    ["prefix", "p"],
-    ["start_offset", "so"],
-    ["streaming_profile", "sp"],
-    ["video_codec", "vc"],
-    ["video_sampling", "vs"],
-  ];
 
-  simple_params.forEach(([name, short]) => {
+  SIMPLE_PARAMS.forEach(([name, short]) => {
     let value = consumeOption(options, name);
     if (value !== undefined) {
       params[short] = value;
@@ -620,13 +510,13 @@ function updateable_resource_params(options, params = {}) {
     params.metadata = utils.encode_context(options.metadata);
   }
   if (options.custom_coordinates != null) {
-    params.custom_coordinates = utils.encode_double_array(options.custom_coordinates);
+    params.custom_coordinates = encodeDoubleArray(options.custom_coordinates);
   }
   if (options.detection != null) {
     params.detection = options.detection;
   }
   if (options.face_coordinates != null) {
-    params.face_coordinates = utils.encode_double_array(options.face_coordinates);
+    params.face_coordinates = encodeDoubleArray(options.face_coordinates);
   }
   if (options.headers != null) {
     params.headers = utils.build_custom_headers(options.headers);
@@ -818,7 +708,7 @@ function video_url(public_id, options) {
 }
 
 function finalize_source(source, format, url_suffix) {
-  var source_to_sign;
+  let source_to_sign;
   source = source.replace(/([^:])\/\//g, '$1/');
   if (source.match(/^https?:\//i)) {
     source = smart_escape(source);
@@ -1412,7 +1302,7 @@ exports.build_upload_params = build_upload_params;
 exports.timestamp = () => Math.floor(new Date().getTime() / 1000);
 exports.option_consume = consumeOption; // for backwards compatibility
 exports.build_array = toArray; // for backwards compatibility
-exports.encode_double_array = encode_double_array;
+exports.encode_double_array = encodeDoubleArray;
 exports.encode_key_value = encode_key_value;
 exports.encode_context = encode_context;
 exports.build_eager = build_eager;
@@ -1453,3 +1343,20 @@ exports.jsonArrayParam = jsonArrayParam;
 // was exported before, so kept for backwards compatibility
 exports.DEFAULT_POSTER_OPTIONS = DEFAULT_POSTER_OPTIONS;
 exports.DEFAULT_VIDEO_SOURCE_TYPES = DEFAULT_VIDEO_SOURCE_TYPES;
+
+Object.assign(module.exports, {
+  at,
+  clone,
+  extend,
+  filter,
+  includes,
+  isArray,
+  isEmpty,
+  isNumber,
+  isObject,
+  isRemoteUrl,
+  isString,
+  isUndefined,
+  keys: source => Object.keys(source),
+  ensurePresenceOf,
+});
