+++ lib/utils/parsing/consumeOption.js
@@ -12,4 +12,4 @@ function consumeOption(options, option_name, default_value) {
   return result != null ? result : default_value;
 }
 
-exports.consumeOption = consumeOption;
+module.exports = consumeOption;
