+++ lib/utils/parsing/toArray.js
@@ -16,4 +16,4 @@ function toArray(arg) {
   }
 }
 
-exports.toArray = toArray;
+module.exports = toArray;
