+++ lib/utils/utf8_encode.js
@@ -20,7 +20,7 @@
  * @return {string}
  */
 module.exports = function utf8_encode(argString) {
-  var c1, enc, n;
+  let c1, enc, n;
   if (argString == null) {
     return "";
   }
