+++ lib/v2/search.js
@@ -64,7 +64,7 @@ const Search = class Search {
   }
 
   sort_by(field_name, dir = "desc") {
-    var sort_bucket;
+    let sort_bucket;
     sort_bucket = {};
     sort_bucket[field_name] = dir;
     this.query_hash.sort_by.push(sort_bucket);
